backend/requirements.txt
asgiref==3.9.1
channels==4.1.0
click==8.2.1
Django==5.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==22.0.0
h11==0.16.0
packaging==25.0
psycopg==3.2.1
psycopg-binary==3.2.1
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
uvicorn[standard]==0.35.0
